feat: implement DAO pattern to object entities
