+++ main.py
@@ -29,32 +29,37 @@ telaAparelho = TelaAparelho(controladorAparelho)
 telaExercicio = TelaExercicio(controladorExercicio, telaAparelho)
 telaProfessor = TelaProfessor(controladorProfessor)
 telaTreino = TelaTreino(controladorTreino, telaAluno, telaExercicio)
-telaHistorico = TelaHistorico(controladorHistorico, telaAluno, telaProfessor, telaTreino)
-telaRelatorios = TelaRelatorios(controladorRelatorios, telaAluno, telaProfessor)
+telaHistorico = TelaHistorico(
+    controladorHistorico, telaAluno, telaProfessor, telaTreino)
+telaRelatorios = TelaRelatorios(
+    controladorRelatorios, telaAluno, telaProfessor)
+
 
 class TelaPrincipal(Tela):
-  def __init__(self, titulo, objeto, opcoes, controlador):
-    super().__init__(titulo, objeto, opcoes, controlador)
+    def __init__(self, titulo, objeto, opcoes, controlador):
+        super().__init__(titulo, objeto, opcoes, controlador)
 
-  def menu_principal(self):
-    telas = {
-      1: ['Tela de Alunos', telaAluno.mostrar_opcoes],
-      2: ['Tela de Professores', telaProfessor.mostrar_opcoes],
-      3: ['Tela de Aparelhos', telaAparelho.mostrar_opcoes],
-      4: ['Tela de Exercícios', telaExercicio.mostrar_opcoes],
-      5: ['Tela de Treinos', telaTreino.mostrar_opcoes],
-      6: ['Tela de Históricos', telaHistorico.mostrar_opcoes],
-      7: ['Tela de Relatórios', telaRelatorios.mostrar_opcoes],
-      8: ['Sair', sys.exit]
-    }
-    print('\n-- CONTROLE DE ACADEMIA --\n')
-    for opcao, tela in telas.items():
-      print(f"{opcao} - {tela[0]}")
-    opcao_escolhida = self.inserir_inteiro('Escolha uma tela: ', list(telas.keys()))
-    telas[opcao_escolhida][1]()
+    def menu_principal(self):
+        telas = {
+            1: ['Tela de Alunos', telaAluno.mostrar_opcoes],
+            2: ['Tela de Professores', telaProfessor.mostrar_opcoes],
+            3: ['Tela de Aparelhos', telaAparelho.mostrar_opcoes],
+            4: ['Tela de Exercícios', telaExercicio.mostrar_opcoes],
+            5: ['Tela de Treinos', telaTreino.mostrar_opcoes],
+            6: ['Tela de Históricos', telaHistorico.mostrar_opcoes],
+            7: ['Tela de Relatórios', telaRelatorios.mostrar_opcoes],
+            8: ['Sair', sys.exit]
+        }
+        print('\n-- CONTROLE DE ACADEMIA --\n')
+        for opcao, tela in telas.items():
+            print(f"{opcao} - {tela[0]}")
+        opcao_escolhida = self.inserir_inteiro(
+            'Escolha uma tela: ', list(telas.keys()))
+        telas[opcao_escolhida][1]()
+
 
 if __name__ == "__main__":
-  rodando = True
-  tela = TelaPrincipal('CONTROLE DE ACADEMIA', None, None, None)
-  while rodando:
-    tela.menu_principal()
+    rodando = True
+    tela = TelaPrincipal('CONTROLE DE ACADEMIA', None, None, None)
+    while rodando:
+        tela.menu_principal()
