+++ modules/pessoa/professor/EntidadeProfessor.py
@@ -1,61 +1,32 @@
 import random
-from errors.IsEmptyError import IsEmptyError
 from modules.pessoa.EntidadePessoa import Pessoa
-from modules.pessoa.professor.turno.EntidadeTurno import Turno
+from modules.pessoa.professor.ProfessorDAO import ProfessorDAO
 
 
-class Professor(Pessoa):
-  table_name = 'Professor'
-  def __init__(self, nome: str, cpf: str, peso: int, altura: float, salario: float, id = random.randint(1000,9999)) -> None:
-    super().__init__(nome, cpf, peso, altura, 'Professor', id)
-    self.__salario = salario
-    self.__turnos_ = None
+class Professor(Pessoa, ProfessorDAO):
 
-  @property
-  def salario(self):
-    return self.__salario
+    def __init__(self, nome: str, cpf: str, peso: int, altura: float, salario: float, id=random.randint(1000, 9999)) -> None:
+        Pessoa.__init__(self, nome, cpf, peso, altura, id)
+        ProfessorDAO.__init__(self)
+        self.__salario = salario
+        self.__turnos_ = None
 
-  @property
-  def turnos(self):
-    return self.__turnos_
-  
-  @turnos.setter
-  def turnos(self, turnos: list):
-    self.__turnos_ = turnos
+    @property
+    def salario(self):
+        return self.__salario
 
-  @salario.setter
-  def salario(self, salario):
-    self.__salario = salario
+    @property
+    def turnos(self):
+        return self.__turnos_
 
-  def criar(self):
-    try:
-      with self.connection:
-        self.cursor.execute(f"""
-          CREATE TABLE IF NOT EXISTS {self.tableName} 
-            (
-              id INTEGER PRIMARY KEY, 
-              nome TEXT, 
-              cpf TEXT, 
-              peso INTEGER,
-              altura REAL,
-              salario TEXT
-            )
-        """)
-      return True
-    except Exception:
-      raise
+    @turnos.setter
+    def turnos(self, turnos: list):
+        self.__turnos_ = turnos
 
-  def buscar_turnos(self) -> list:
-    try:
-      res = Professor.cursor.execute(f"SELECT * FROM {Turno.table_name} WHERE professor = {self.identificador}")
-      return [dict(row) for row in res.fetchall()]
-    except:
-      raise IsEmptyError
+    @salario.setter
+    def salario(self, salario):
+        self.__salario = salario
 
-  @staticmethod
-  def buscar() -> list:
-    try:
-      res = Professor.cursor.execute(f"SELECT * FROM {Professor.table_name}")
-      return [dict(row) for row in res.fetchall()]
-    except:
-      raise IsEmptyError
+    @staticmethod
+    def buscar() -> list:
+        return ProfessorDAO.buscar()
