+++ modules/treino/EntidadeTreino.py
@@ -1,81 +1,47 @@
 import random
-from abstract.entidade import Entidade
-from errors.IsEmptyError import IsEmptyError
 from modules.pessoa.aluno.EntidadeAluno import Aluno
+from modules.treino.TreinoDAO import TreinoDAO
 
-class Treino(Entidade):
-  table_name = 'Treino'
-  def __init__(self, nome: str, aluno: Aluno, id = None) -> None:
-    if not id: id = random.randint(1000,9999)
-    super().__init__('Treino', 'id')
-    self.__id = id
-    self.__nome = nome
-    self.__aluno = aluno
-    self.__praticas_ = None
 
-  @property
-  def identificador(self):
-    return self.__id
+class Treino(TreinoDAO):
 
-  @property
-  def nome(self):
-    return self.__nome
+    def __init__(self, nome: str, aluno: Aluno, id=None) -> None:
+        if not id:
+            id = random.randint(1000, 9999)
+        super().__init__()
+        self.__id = id
+        self.__nome = nome
+        self.__aluno = aluno
+        self.__praticas_ = None
 
-  @nome.setter
-  def nome(self, nome):
-    self.__nome = nome
-   
-  @property
-  def aluno(self):
-    return self.__aluno
+    @property
+    def identificador(self):
+        return self.__id
 
-  @aluno.setter
-  def aluno(self, aluno):
-    self.__aluno = aluno
-  
-  @property
-  def praticas(self):
-    return self.__praticas_
-  
-  @praticas.setter
-  def praticas(self, praticas: list):
-    self.__praticas_ = praticas
+    @property
+    def nome(self):
+        return self.__nome
 
-  def criar(self):
-    try:
-      with self.connection:
-        self.cursor.execute(f"""
-          CREATE TABLE IF NOT EXISTS {self.tableName} 
-            (id INTEGER PRIMARY KEY, nome TEXT, aluno INTEGER NOT NULL,
-             FOREIGN KEY (aluno) REFERENCES Aluno (id))
-        """)
-      return True
-    except Exception:
-      raise
+    @nome.setter
+    def nome(self, nome):
+        self.__nome = nome
 
-  def buscar_praticas(self) -> list:
-    try:
-      res = Treino.cursor.execute(f'''
-        SELECT 
-          Pratica.id,
-          Pratica.repeticoes,
-          Pratica.peso,
-          Pratica.treino,
-          Exercicio.nome as exercicio
-        FROM
-          Pratica
-          INNER JOIN Exercicio ON Exercicio.id = Pratica.exercicio
-        WHERE 
-          treino = {self.identificador};
-      ''')
-      return [dict(row) for row in res.fetchall()]
-    except:
-      raise IsEmptyError
+    @property
+    def aluno(self):
+        return self.__aluno
 
-  @staticmethod
-  def buscar() -> list:
-    try:
-      res = Treino.cursor.execute(f"SELECT * FROM {Treino.table_name}")
-      return [dict(row) for row in res.fetchall()]
-    except:
-      raise IsEmptyError
+    @aluno.setter
+    def aluno(self, aluno):
+        self.__aluno = aluno
+
+    @property
+    def praticas(self):
+        return self.__praticas_
+
+    @praticas.setter
+    def praticas(self, praticas: list):
+        self.__praticas_ = praticas
+
+    @staticmethod
+    def buscar() -> list:
+        return TreinoDAO.buscar()
